LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-null-or-undefined-expected.txt


PASS event.intercept() should throw if the handler is null
FAIL event.intercept() should not throw if the handler is explicit undefined assert_equals: expected "#2" but got "#1"
